Refactor cmd/utils.go
