+++ cmd/dsn.go
@@ -1,0 +1,3 @@
+package cmd
+
+const DSN = "https://e22f42dbb3e0433fbd9ea32453faa598@o4504550475038720.ingest.sentry.io/4504550481723392"
