+++ cmd/plugin_list.go
@@ -1,6 +1,9 @@
 package cmd
 
 import (
+	"context"
+	"strings"
+
 	"github.com/gatewayd-io/gatewayd/config"
 	"github.com/getsentry/sentry-go"
 	"github.com/spf13/cobra"
@@ -49,4 +52,35 @@ func init() {
 		false, "Only list enabled plugins")
 	pluginListCmd.Flags().BoolVar(
 		&enableSentry, "sentry", true, "Enable Sentry") // Already exists in run.go
+}
+
+func listPlugins(cmd *cobra.Command, pluginConfigFile string, onlyEnabled bool) {
+	// Load the plugin config file.
+	conf := config.NewConfig(context.TODO(), "", pluginConfigFile)
+	conf.LoadDefaults(context.TODO())
+	conf.LoadPluginConfigFile(context.TODO())
+	conf.UnmarshalPluginConfig(context.TODO())
+
+	if len(conf.Plugin.Plugins) != 0 {
+		cmd.Printf("Total plugins: %d\n", len(conf.Plugin.Plugins))
+		cmd.Println("Plugins:")
+	} else {
+		cmd.Println("No plugins found")
+	}
+
+	// Print the list of plugins.
+	for _, plugin := range conf.Plugin.Plugins {
+		if onlyEnabled && !plugin.Enabled {
+			continue
+		}
+		cmd.Printf("  Name: %s\n", plugin.Name)
+		cmd.Printf("  Enabled: %t\n", plugin.Enabled)
+		cmd.Printf("  Path: %s\n", plugin.LocalPath)
+		cmd.Printf("  Args: %s\n", strings.Join(plugin.Args, " "))
+		cmd.Println("  Env:")
+		for _, env := range plugin.Env {
+			cmd.Printf("    %s\n", env)
+		}
+		cmd.Printf("  Checksum: %s\n", plugin.Checksum)
+	}
 }
